requirements.txt
numpy
matplotlib
pillow
pyrealsense2
